pyproject.toml
[project]
name = "rag_project"
version = "0.1.0"
description = "Modular RAG console application"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "langchain",
    "langchain-core",
    "langchain-community",
    "langchain-text-splitters",
    "google-genai",
    "chromadb",
    "pydantic",
    "pyyaml",
    "python-dotenv",
    "pymupdf",
    "sentence-transformers",
    "numpy",
    "tenacity",
]

# Note: PyTorch should be installed separately based on your hardware.
# See README.md for platform-specific installation instructions.

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest",
]

